chore: move lib/client to features
