+++ packages/react-server/src/client.tsx
@@ -1,7 +1,7 @@
 "use client";
 
-export { Link, LinkForm } from "./lib/client/link";
-export { useRouter } from "./lib/client/router";
+export { Link, LinkForm } from "./features/router/client/link";
+export { useRouter } from "./features/router/client/router";
 export {
   routerRevalidate,
   useParams,
