+++ packages/react-server/src/entry-browser.tsx
@@ -12,17 +12,17 @@ import {
   routerRevalidate,
 } from "./features/router/client";
 import {
+  Router,
+  RouterContext,
+  createEncodedBrowserHistory,
+  useRouter,
+} from "./features/router/client/router";
+import {
   type RouteManifest,
   emptyRouteManifest,
 } from "./features/router/manifest";
 import type { ServerRouterData } from "./features/router/utils";
 import { createStreamRequest } from "./features/server-component/utils";
-import {
-  Router,
-  RouterContext,
-  createEncodedBrowserHistory,
-  useRouter,
-} from "./lib/client/router";
 import { $__global } from "./lib/global";
 import type { CallServerCallback } from "./lib/types";
 import { getFlightStreamBrowser } from "./utils/stream-script";
